src/requirements.txt
asyncio
datalake-scripts
python-dotenv
requests
schedule
wheel
urllib3